--- a/pkg/streamingpromql/testdata/upstream/aggregators.test
+++ b/pkg/streamingpromql/testdata/upstream/aggregators.test
@@ -60,12 +60,11 @@ eval instant at 50m sum without () (http_requests{job="api-server",group="produc
   {group="production",job="api-server",instance="1"} 200
 
 # Without with mismatched and missing labels. Do not do this.
-# Unsupported by streaming engine.
-# eval instant at 50m sum without (instance) (http_requests{job="api-server"} or foo)
-#   {group="canary",job="api-server"} 700
-#   {group="production",job="api-server"} 300
-#   {region="europe",job="api-server"} 900
-#   {job="api-server"} 1000
+eval instant at 50m sum without (instance) (http_requests{job="api-server"} or foo)
+  {group="canary",job="api-server"} 700
+  {group="production",job="api-server"} 300
+  {region="europe",job="api-server"} 900
+  {job="api-server"} 1000
 
 # Lower-cased aggregation operators should work too.
 eval instant at 50m sum(http_requests) by (job) + min(http_requests) by (job) + max(http_requests) by (job) + avg(http_requests) by (job)