--- a/pkg/streamingpromql/testdata/upstream/range_queries.test
+++ b/pkg/streamingpromql/testdata/upstream/range_queries.test
@@ -62,10 +62,9 @@ load 30s
   foo{job="1"} 1+1x4
   bar{job="2"} 1+1x4
 
-# Unsupported by streaming engine.
-# eval range from 0 to 2m step 1m foo > 2 or bar
-#   foo{job="1"} _ 3 5
-#   bar{job="2"} 1 3 5
+eval range from 0 to 2m step 1m foo > 2 or bar
+  foo{job="1"} _ 3 5
+  bar{job="2"} 1 3 5
 
 clear