… CHANGELOG.md
… pkg/streamingpromql/operators/or_binary_operation_test.go
… pkg/streamingpromql/testdata/ours/binary_operators.test